+++ lib_xua/src/dfu/xua_dfu.h
@@ -13,10 +13,6 @@
 #error DFU_PID not defined!
 #endif
 
-#ifndef DFU_BCD_DEVICE
-#error DFU_BCD_DEVICE not defined!
-#endif
-
 #ifndef DFU_SERIAL_STR_INDEX
 /* By default no serial string */
 #define DFU_SERIAL_STR_INDEX        (0x00)
@@ -43,8 +39,8 @@ unsigned char DFUdevDesc[] = {
     (DFU_VENDOR_ID >> 8),           /* 9  idVendor */
     (DFU_PID & 0xFF),               /* 10 idProduct */
     (DFU_PID >> 8),                 /* 11 idProduct */
-    (DFU_BCD_DEVICE & 0xFF),        /* 12 bcdDevice : Device release number */
-    (DFU_BCD_DEVICE >> 8),          /* 13 bcdDevice : Device release number */
+    (BCD_DEVICE & 0xFF),        /* 12 bcdDevice : Device release number */
+    (BCD_DEVICE >> 8),          /* 13 bcdDevice : Device release number */
     DFU_MANUFACTURER_STR_INDEX,     /* 14 iManufacturer : Index of manufacturer string */
     DFU_PRODUCT_STR_INDEX,          /* 15 iProduct : Index of product string descriptor */
     DFU_SERIAL_STR_INDEX,           /* 16 iSerialNumber : Index of serial number decriptor */